Add a module to initialize the database.
Testing this involved consolidating code that spins up the
database, so we can share a database instance.

I also pulled code that creates database client pools outwards
so that the test code can configure it to not cause test failure
when a client in the pool detects that the server has shutdown
hard.
This wouldn't be necessary if Pool.end(callback) actually invoked
the callback, but for some reason it does not, so all calls to
Pool.end are fire-and-forget.