Rollback to last version that found postgres binary.
Recent travis builds have failed on `which postgres` in scripts/run-locally.js.

This rolls back to github.com/mikesamuel/attack-review-testbed/blob/8a922e9327fe7dc2b38391d3a647cf270acc5c32/.travis.yml
which seemed to fail due to use of `IF NOT EXISTS` syntax in lib/db-tables.js.

https://travis-ci.org/mikesamuel/attack-review-testbed/jobs/438755825 says

> ```
> 1) db-tables
>       createTables:
>     error: syntax error at or near "NOT"
> ```

which seems to indicate that either CREATE TABLE __ IF NOT EXISTS or
CREATE INDEX __ IF NOT EXISTS fails on the version of postgres provided
by

```
services
  - postresql
```

Trying to use addons and apt-get to install a more recent version fail to put
a postgres executable where its needed.