lockdown builtin "vm" module
Derived a whitelist of modules that use the "vm" module.
Now the vast majority of loaded modules cannot use vm to load code from
strings.

I derived this whitelist by adding

```json
  "vm": {
    "mode": "report-only",
    "ids": []
  }
```

to the sensitiveModules property in package.json, running it and looking at the trace.

I declined to whitelist jsdom related modules on the grounds that I don't want it
loading code.